test: consolidate unit tests for non single field partitioning dtos
to reduce the number of unit test files

apache/gravitino#5199
